refactor: change a few things around
