+++ netlify.toml
@@ -6,4 +6,4 @@
   force = true
   from = "/*"
   status = 200
-  to = "/.netlify/functions/app/:splat"
+  to = "/.netlify/functions/server-prod/:splat"
