+++ package.json
@@ -4,7 +4,7 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "node app.js",
+    "dev": "node server-dev.js",
     "build": "node build"
   },
   "dependencies": {
