+++ src/pages/index/function.js
@@ -1,4 +1,4 @@
-export const GET = async () => {
+export const getServerData = async () => {
   const request = await fetch('https://dummyjson.com/products?limit=10');
   const data = await request.json();
