+++ src/pages/page-2/function.js
@@ -1,4 +1,4 @@
-export const GET = async () => {
+export const getServerData = async () => {
   const request = await fetch('https://dummyjson.com/carts?limit=10');
   const data = await request.json();
